LDE-714: moved tests for cleanup_cube
